requirements.txt
streamlit>=1.30.0
folium>=0.15.0
streamlit-folium>=0.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
lxml>=5.0.0
PyYAML>=6.0
